Renamed the package xmlModelParser to xmlParser
Renamed the package xmlModelParser to xmlParser
