+++ Project7/src/main/java/models/Config.java
@@ -5,12 +5,12 @@ import java.util.Observable;
 import org.w3c.dom.Element;
 import org.w3c.dom.NodeList;
 
-import xmlModelParser.Parsable;
-import xmlModelParser.ParserException;
-import xmlModelParser.XmlAtribute;
-import xmlModelParser.XmlList;
-import xmlModelParser.XmlNode;
-import xmlModelParser.XmlString;
+import xmlParser.Parsable;
+import xmlParser.ParserException;
+import xmlParser.XmlAtribute;
+import xmlParser.XmlList;
+import xmlParser.XmlNode;
+import xmlParser.XmlString;
 
 /**
  * Represents the XML tag "config", implements {@link Parsable}.
