+++ Project7/src/main/java/models/Test.java
@@ -4,11 +4,11 @@ import java.util.Observable;
 
 import org.w3c.dom.Element;
 
-import xmlModelParser.Parsable;
-import xmlModelParser.ParserException;
-import xmlModelParser.XmlAtribute;
-import xmlModelParser.XmlNode;
-import xmlModelParser.XmlString;
+import xmlParser.Parsable;
+import xmlParser.ParserException;
+import xmlParser.XmlAtribute;
+import xmlParser.XmlNode;
+import xmlParser.XmlString;
 
 /**
  * Represents the XML tag "test", implements {@link Parsable}.
