+++ Project7/src/main/java/xmlParser/Parsable.java
@@ -1,4 +1,4 @@
-package xmlModelParser;
+package xmlParser;
 
 import java.util.Observer;
