+++ Project7/src/main/java/xmlParser/Parser.java
@@ -1,4 +1,4 @@
-package xmlModelParser;
+package xmlParser;
 
 import java.io.BufferedWriter;
 import java.io.File;
