+++ Project7/src/main/java/xmlParser/ParserException.java
@@ -1,4 +1,4 @@
-package xmlModelParser;
+package xmlParser;
 
 public class ParserException extends Exception {
