+++ Project7/src/main/java/xmlParser/XmlList.java
@@ -1,4 +1,4 @@
-package xmlModelParser;
+package xmlParser;
 
 import java.util.ArrayList;
